Pull out ga() calls into utils
